Fix thdid bug
Fixed issue with the definition of sl_thd in rust that was causing the thread
id to get lost. Changed sl_thd to match c definition and added necessary
composite types. Changed sl::Thread to just contain the thread id and changed
sl functions that previously reference the thd_ptr to now just use the thdid or
directly calling out to sl.